Added further fallback implementation and fixed a few bugs.

+++ eos/bootstrap.py
@@ -11,7 +11,7 @@ import eos.util
 
 
 def bootstrap_library(json_obj, name, library_dir, postprocessing_dir, create_snapshots=False,
-                      fallback_server_url=None):
+                      fallback_server_url=None, force_fallback=False):
     eos.log("Bootstrapping library '" + name + "' to " + library_dir)
 
     # create directory for library
@@ -36,12 +36,14 @@ def bootstrap_library(json_obj, name, library_dir, postprocessing_dir, create_sn
         eos.log_warning("unknown source type for library '" + name)
         return False
 
-    def get_from_fallback(filename, download_dir):
+    def get_from_fallback(filename, relative_src_dir, download_dir):
         if fallback_server_url is None:
             return False
         eos.log("downloading repository from fallback URL %s..." % fallback_server_url)
         fallback_success = eos.fallback.download_and_extract_from_fallback_url(fallback_server_url, filename,
-                                                                               download_dir, library_dir)
+                                                                               relative_src_dir=relative_src_dir,
+                                                                               download_dir=download_dir,
+                                                                               extract_dir=library_dir)
         if not fallback_success:
             eos.log_error("download from fallback URL failed")
         return fallback_success
@@ -51,11 +53,16 @@ def bootstrap_library(json_obj, name, library_dir, postprocessing_dir, create_sn
         sha1_hash = src.get('sha1', None)
         user_agent = src.get('user-agent', None)
 
+        if force_fallback:
+            return get_from_fallback(eos.util.get_filename_from_url(eos.util.sanitize_url(src_url)),
+                                     eos.cache.get_relative_archive_dir(), eos.cache.get_archive_dir())
+
         # download archive file
         download_filename = eos.util.download_file(src_url, eos.cache.get_archive_dir(), sha1_hash, user_agent)
         if download_filename == "":
             eos.log_error("downloading of file for '" + name + "' from " + src_url + " failed")
-            return get_from_fallback(os.path.basename(download_filename), eos.cache.get_archive_dir())
+            return get_from_fallback(os.path.basename(download_filename),
+                                     eos.cache.get_relative_archive_dir(), eos.cache.get_archive_dir())
 
         if os.path.exists(library_dir):
             shutil.rmtree(library_dir)
@@ -63,7 +70,8 @@ def bootstrap_library(json_obj, name, library_dir, postprocessing_dir, create_sn
         # extract archive file
         if not eos.archive.extract_file(download_filename, library_dir):
             eos.log_error("extraction of file for '" + download_filename + "' failed")
-            return get_from_fallback(os.path.basename(download_filename), eos.cache.get_archive_dir())
+            return get_from_fallback(os.path.basename(download_filename),
+                                     eos.cache.get_relative_archive_dir(), eos.cache.get_archive_dir())
     else:
         # We're dealing with a repository
         branch = src.get('branch', None)
@@ -80,9 +88,13 @@ def bootstrap_library(json_obj, name, library_dir, postprocessing_dir, create_sn
             snapshot_archive_name = name + "_" + revision + ".tar.gz"  # add the revision number, if present
 
         # clone or update repository
-        if not eos.repo.update_state(src_type, src_url, name, library_dir, branch, revision):
-            eos.log_error("updating repository state for '" + name + " failed")
-            fallback_success = get_from_fallback(snapshot_archive_name, eos.cache.get_snapshot_dir())
+        if not force_fallback:
+            update_repo_success = eos.repo.update_state(src_type, src_url, name, library_dir, branch, revision)
+        if force_fallback or not update_repo_success:
+            if not force_fallback:
+                eos.log_error("updating repository state for '" + name + " failed")
+            fallback_success = get_from_fallback(snapshot_archive_name,
+                                                 eos.cache.get_relative_snapshot_dir(), eos.cache.get_snapshot_dir())
             if not fallback_success:
                 return False
             fallback_success = eos.repo.update_state(src_type, None, name, library_dir, branch, revision)
